feat: support SequentialAnimate code load by manual
