+++ packages/vchart/src/animation/index.ts
@@ -1,0 +1,3 @@
+export { registerAnimate, registerCustomAnimate } from '@visactor/vrender-animate';
+export { registerStateTransition } from './state-transition';
+export { registerSequentialAnimate } from './sequential-animate';
